requirements.txt
requests>=2.31.0
llama-cpp-python>=0.2.82
mlx-lm>=0.10.0; platform_system == "Darwin"
